changed udelay
git-svn-id: https://trigonak-thesis.googlecode.com/svn/exjobb/iRCCEdeddsl@2059 dd814760-9b98-ad33-1d6a-b699670a4318

+++ include/tm.h
@@ -137,7 +137,7 @@ int comsz; RCCE_comm_size(RCCE_COMM_APP, &comsz);PRINTD("size of comm_app is %d"
 
     inline double udelay(unsigned int micros) {
         double __ts_start = RCCE_wtime();
-        double __ts_end = __ts_start + (micros / 1000000);
+        double __ts_end = __ts_start + ((double) micros / 1000000);
         PRINTD("Start: %f, End: %f", __ts_start, __ts_end);
         while (RCCE_wtime() < __ts_end);
         return __ts_start - __ts_end;
